+++ src/contexts/plants/application/services/read/enrich-plant-with-qr/enrich-plant-with-qr.service.spec.ts
@@ -45,7 +45,11 @@ describe('EnrichPlantWithQrService', () => {
   let plantBuilder: PlantBuilder;
 
   beforeEach(() => {
-    qrPort = { findByQrId: jest.fn() };
+    qrPort = {
+      findByQrId: jest.fn(),
+      createForPlant: jest.fn(),
+      delete: jest.fn(),
+    };
     plantBuilder = new PlantBuilder();
     service = new EnrichPlantWithQrService(plantBuilder, qrPort);
   });